@@ -4,32 +4,34 @@
   "description": "> Project description",
   "main": "index.js",
   "dependencies": {
-    "@babel/core": "^7.12.10",
     "@babel/node": "^7.12.10",
     "@babel/preset-env": "^7.12.11",
     "@babel/preset-react": "^7.12.10",
-    "babel-loader": "^8.2.2",
+    "axios": "^0.21.1",
     "body-parser": "^1.19.0",
     "express": "^4.17.1",
     "faker": "^5.1.0",
     "init": "^0.1.2",
     "mongodb": "^3.6.3",
     "mongoose": "^5.11.11",
+    "path": "^0.12.7",
     "react": "^17.0.1",
     "react-dom": "^17.0.1",
-    "webpack": "^5.14.0",
     "webpack-cli": "^4.3.1"
   },
   "devDependencies": {
+    "@babel/core": "^7.12.10",
     "babel-eslint": "^10.1.0",
     "babel-jest": "^26.6.3",
+    "babel-loader": "^8.2.2",
     "eslint": "^7.17.0",
+    "eslint-config-hackreactor": "git://github.com/reactorcore/eslint-config-hackreactor",
     "eslint-plugin-react": "^7.22.0",
     "eslint-plugin-react-hooks": "^4.2.0",
     "jest": "^26.6.3",
     "nodemon": "^2.0.7",
     "react-test-renderer": "^17.0.1",
-    "eslint-config-hackreactor": "git://github.com/reactorcore/eslint-config-hackreactor"
+    "webpack": "^5.15.0"
   },
   "scripts": {
     "test": "jest",
